+++ src/console/controllers/RefreshController.php
@@ -21,6 +21,12 @@ class RefreshController extends Controller
     public $volume;
 
     /**
+     * @var string|null
+     * @since 2.10.0
+     */
+    public $provider;
+
+    /**
      * @inheritdoc
      */
     public function options($actionID)
@@ -31,12 +37,16 @@ class RefreshController extends Controller
             $options[] = 'volume';
         }
 
+        if ($actionID === 'by-provider') {
+            $options[] = 'provider';
+        }
+
         return $options;
     }
 
     public function actionAll(): int
     {
-        return $this->_refresh(null);
+        return $this->_refresh(null, null);
     }
 
     public function actionByVolume(): int
@@ -46,10 +56,26 @@ class RefreshController extends Controller
             return ExitCode::UNSPECIFIED_ERROR;
         }
 
-        return $this->_refresh($this->volume);
+        return $this->_refresh(explode(',', $this->volume), null);
     }
 
-    private function _refresh(?string $volume): int
+    /**
+     * Refreshes embedded assets by provider. The provider(s) must match the `providerName` of the embedded assets.
+     *
+     * @return int
+     * @since 2.10.0
+     */
+    public function actionByProvider(): int
+    {
+        if ($this->provider === null) {
+            $this->stderr('The --provider option must be specified with the by-provider action.' . PHP_EOL, Console::FG_RED);
+            return ExitCode::UNSPECIFIED_ERROR;
+        }
+
+        return $this->_refresh(null, explode(',', $this->provider));
+    }
+
+    private function _refresh(?array $volume, ?array $providers): int
     {
         $assetsService = Craft::$app->getAssets();
         $elementsService = Craft::$app->getElements();
@@ -59,11 +85,21 @@ class RefreshController extends Controller
         $assets = Asset::find()->kind('json');
 
         if ($volume !== null) {
-            $assets->volume = explode(',', $volume);
+            $assets->volume($volume);
+        }
+
+        $providersKeys = [];
+
+        if ($providers !== null) {
+            foreach ($providers as $provider) {
+                $providersKeys[$provider] = true;
+            }
         }
 
         foreach ($assets->all() as $asset) {
-            if (($embeddedAsset = EmbeddedAssets::$plugin->methods->getEmbeddedAsset($asset)) !== null) {
+            $embeddedAsset = EmbeddedAssets::$plugin->methods->getEmbeddedAsset($asset);
+
+            if ($embeddedAsset !== null && ($providers === null || isset($providersKeys[$embeddedAsset->providerName]))) {
                 $embeddedAssets[$asset->id] = [
                     'asset' => $asset,
                     'embeddedAsset' => $embeddedAsset,